sync drawId to uint24
sync timestamps to uint48, & bump fuzz prize amounts to uint88 since max uint64 is only 18.4 prize tokens

+++ src/libraries/TierCalculationLib.sol
@@ -81,7 +81,7 @@ library TierCalculationLib {
   /// @param _winningRandomNumber The winning random number
   /// @return A pseudo-random number
   function calculatePseudoRandomNumber(
-    uint32 _drawId,
+    uint24 _drawId,
     address _vault,
     address _user,
     uint8 _tier,
